#4 Changes to background image
Conversion of background image to b&w created some unattractive artifacts, especially around the head of one of the main figures as noted by @jvasile. New image has been converted in a different way and image was also changed. Result is smoother and also not as dark. As a small cosmetic tweak, some bottom padding was also added to the div containing the user instructions.