#7: Document couch_views.js
We're doing queries the CouchDB way, with views.  These are functions that
return JSON results from the database.  I've added a lot more information about
them -- it would be overkill for a couch expert, but since most of us are
beginners I think it'll be helpful.  Some of this information (especially that
in DATA.md) might need to be transferred to INSTALL.md at some point.