Use differently-sized input fields (issue #17).
* views/index.jade: Use different Bootstrap widths for different
  fields.

This is actually a bit of a cargo-cultish way to accomplish the goal,
since the resultant widths still aren't always a perfect match for
their inputs, but things are a lot closer now.  I'll hold off on
further tweaks of this nature, though, because (as per issue #23)
we're going to revamp the front page anyway.