#8 Import styles from DCSOps
Font on our site is now Helvetica Neue, matching that on DCSOps. Border radii on divs are now the same and the colors of the submit button now match DCSOps buttons in all states. I've also grabbed the style sheet from DCSOps so its rules will be applied to other elements that the two websites have in common. All rules can still be overridden in "base.css", however. For example our font sizes are generally larger than those on DCSOps, which pertains to #21.

That said, our site and www.dcsops.org are still very different looking websites. However, elements that are common to both should be styled identically.