Remove "Permission to text" paradigm.
This commit is part of issue #23.

Making this change because @nttaylor is revamping the front-page UI
anyway, as per issue #15, and will completely redo how we ask about
text vs phone vs email.  In the meantime, the existing label and
button were adding a lot of visual clutter without being useful.

Soon we'll be presenting this choice to the user as a set of options
("Would you like us to text, phone, or email you?"), and whatever they
choose will implicitly grant us permission to use that method(s) to
contact them.  Plus, we'll have a privacy policy anyway, as per issue
#9, which further obviates the need for any kind of "permission to do
X" question on the front page.

Leaving issue #26 open because when we have checkboxes the point
raised in that issue will still apply to the new UI elements.

By the way, note one thing we *can't* do is automatically check if a
phone number is textable, so we'll definitely have to ask the user.
See [Twilio's FAQ
this](https://www.twilio.com/help/faq/sms/does-twilio-check-to-see-if-phone-numbers-can-receive-sms)
for more.