#22
This commit adds red asterisks next to required form fields which can be changed to warning messages when form is submitted incomplete. I should have been doing this work on the master branch instead of here, since this branch was originally to tackle #15, and I had already begun work on that issue when the ideas for the design were slightly changed. Now this revision is in a limbo state of partial completion for #15 and #22. I'm committing now so I can get back on the main branch and implement changes for #22. Then I will merge those changes back over here and use this branch only for #15.