#22 Errors now more obvious, yet less disruptive
Red asterisks now mark required form fields. If validation script
returns error messages, these now appear in place of the asterisks and
the errored fields gain a red outline. I have also commented out the
link to the DCSOps stylesheet, as one of the rules was breaking the
behavior of the bootstrap grid especially for extra-small screen sizes
but in other ways also.