Removed unnecessary Ruby script used for transforming data
At one point, the data we had received had had just state initials in
its database. Since we needed to have full states in the data, I wrote
a simple script to make that substitution. While it was helpful,
considering that the substitution can be done in many ways, I figured
that it could be removed from the application.