Updated test for zipcode not found to fulfill XPASS condition
There are situations where zipcodes could seem to be valid, but do not
actually exist. In these cases, the sorry page includes some logic
for that eventuality. This ensures that the test the corresponds
to that outcome has a valid zip code for this test (which is an
invalid zipcode)