#15 & #22 Fix checkbox functionality
Necessary name changes in certain DOM elements that were committed in the last revision are now reflected in all .js files that reference those elements. This corrects errors in both appearance and functionality introduced in the last revision. Additional third-party files (font-awesome) are now stored locally to reduce reliance on network connection.

This commit is from @nttaylor, not me!  He is offline and can't ssh to GitHub.