Conditionalize debugging output (issue #43).
Committing on the `issue-43-experimental` branch because this solution
is not ideal and should be considered experimental.  It makes a
conditional-logging function available via the `db` object, but `db`
is an odd place for that.  In IRC @trystant recommended winston: "we
can use winston to set default logging levels per environment and then
route each statement to the appropriate logging level; info, debug,
etc.  https://www.npmjs.com/package/winston#logging-levels".

* models/index.js: Convert config.logging value.
  (maybe_console_log): New exported function.

* routes/index.js: Use maybe_console_log() everywhere.

* config/config.json.tmpl: Update comment about logging.

* app.js: Be silent if DB sync succeeds, error if it fails.