#54: Make site vertically oriented
Move text to above and below fields (instead of left and right), and
widen fields slightly.  Make most text smaller, except for an
introductory sentence which is now bigger.  I made the text smaller so
that we wouldn't need to scroll, but it might now be too small.

I altered views/layout.jade to remove the navigation div at the top,
also in the interest of avoiding scrolling as much as possible.

I'm not sure if I made the layout changes in the correct way, and would
welcome some commit review from @nttaylor or @msnyon.