#54 Remove red stars; tweak text
Removes the red stars that had been to the right of required fields
because all fields are now required. The text "please fill in all
fields..." is hopefully enough of an indication of this. The <div>s
where error messages appear still (now incongruously) belong to class
"warning-star". This might go away soon pending a new implementation for
how warning messages are displayed.

Also change "10 year" to "10-year" because this is a compound adjective
in attributive position. And if you don't know what that means, you can
read about it here: http://www.grammar.com/compound-adjectives/