#46 Continue work on better validation messages
Validation messages are in place for all required fields (which is all
fields). Further tweaks to position may be necessary as we experiment
with changes in the positions, functions, and number of the form fields.
The addition of a regex to the English locale file allows validation on
US-style 10-digit telephone numbers.

For the moment, files form the previous validation library remain in the
repository, leaving a clutter of files in the 'third-party' directory
with 'validate' or 'validation' in the name. If we go forward with this
solution, files from the old library should be removed, along with
<div>s of class 'warning-star' (in 'index.jade'), which had previously
marked recepticles for validation error messages, which would no longer
be used for anything.