More work on new SQL-based data import process.
This commit fixes some of the issues with commit fe99582e on `master`,
but is still not working.

Commit fe99582e wasn't yet working either (see below); this commit may
get us a bit closer, although it is still broken.  It fills in some
(perhaps all) of the missing documentation, fixes some bugs in the new
JSON->SQL conversion script, and attempts to add the missing step of
table creation (though perhaps not correctly -- see details below).

Here are the changes:

* INSTALL.md: Document new import procedures and npm requirements.

* data/create_static_tables.sql: New file.

* data/import_into_postgres.js: Rename to...

* data/json_to_sql.js: ...here, to reflect the functionality change
  implemented in commit fe99582e.  Update documentation.  Correct
  output table name from "UsAddress" to "UsAddresses".  Correct
  column name from "country" to "county", so that we don't end up with
  every county in the database stored as "US" County :-).  Remove the
  SQL files before recreating them, to avoid cross-run appending.

* data/selected_counties.sql: Remove; see commit e46c8c198 for why.

* data/.gitignore: Add this, to ignore .sql files here.

After this commit, `npm start` still shows errors.  The errors have
been inconsistent, so I'm not bothering to transcribe them here.

However, if one does the data conversion and import, and then switches
to the `email_demo` branch and runs `npm start` there, the application
will start normally.  The next problem comes when one fills in the web
form and hits Submit -- we get an error indicating, probably, that
something is wrong in how `data/create_static_tables.sql` sets up the
sequence tables and/or primary key.

Here is that error, which comes right after the debug output showing
that the email has been sent.  This is from right after hitting
Submit, after `npm start` on the `email_demo` branch (but having
converted and loaded data using the new method on this branch):

    DEBUG: This is the email we're about to send:

    (...details of email here, as expected...)

    DEBUG: (end of email)

    Unhandled rejection TypeError: Cannot read property 'id' of undefined
        at /home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/routes/index.js:300:29
        at rejectError (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/mailgun-js/lib/build.js:108:3)
        at constructor.(anonymous function) [as send] (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/mailgun-js/lib/build.js:75:14)
        at null.<anonymous> (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/routes/index.js:297:39)
        at tryCatcher (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/sequelize/node_modules/bluebird/js/main/util.js:26:23)
        at Promise._settlePromiseFromHandler (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/sequelize/node_modules/bluebird/js/main/promise.js:503:31)
        at Promise._settlePromiseAt (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/sequelize/node_modules/bluebird/js/main/promise.js:577:18)
        at Async._drainQueue (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/sequelize/node_modules/bluebird/js/main/async.js:128:12)
        at Async._drainQueues (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/sequelize/node_modules/bluebird/js/main/async.js:133:10)
        at Async.drainQueues (/home/kfogel/private/work/ots/client/red-cross/r/smoke-alarm-portal/node_modules/sequelize/node_modules/bluebird/js/main/async.js:15:14)
        at process._tickCallback (node.js:419:13)