Sequelize Import 2/3 - Import works, I Promise
The promise chain is in place to make sure that the appropriate
tables are created before putting data in each of them.

Also logging is set to false, which should remove the logging
but does not seem to work. I need to open a ticket with sequelize

The same command (node data/import_into_postgres.js) will work
but will take several minutes.