Make a "/admin" area with very basic Basic Auth.
This `simple_admin` branch is just an experiment.  The work for #7 is
on `master` and is based on Drywall, which is a much better foundation
for building long-term extensible admin functionality.  However, that
better foundation may also need a bit of baking time; meanwhile, we
want to be able to do iterative UI development with the client, which
this branch may enable cheaply and lightweightly (one of those things
isn't an adverb) enough to avoid delays in getting feedback.

The admin area here is just a hello-world page so far.  It doesn't
have any functionality, such as reports, yet.

* package.json: Require "basic-auth".

* views/admin.jade: New file.

* routes/index.js: Add "/admin" route.