Request filtering working: # 82 part 2
The search, querying and filtering logic was not set to work
with Sequelize. The pagedFinder in Drywall was a Mongoose plugin
whose code I took, modified and put in the getResults() method
of the admin index page. I then modified the front page to match
the results as expected within drywall

As a result, the filters now shows the results. Paging should be
in as well, but there's a bug to resolve with the paging view showing
up. I do believe a small tweak will fix it, however