Reverse order of steps 6 and 7 in INSTALL.md.
In theory, it should be possible to do the data import before running
`npm install`.  But right now we have a kludge whereby some of the
mailgun setup stuff is hooked on the db object in `models/index.js`,
which means that now doing a data import (thus using db) has an
unexpected dependency on mailgun.  This is highly un-ideal, but for
now just requiring the `npm install` step (to get mailgun) before
importing any data is an acceptable workaround.