Require `moment` in package.json.
This shouldn't be necessary, because in theory, we've had `moment` in
the tree in `public/vendor/momentjs/` since commit bfcd43f5.  But
actually, when app.js requires `moment`, it's been getting it from
`node_modules/` all along -- and we just always happened to have it
installed there (from having fetched it manually during development,
presumably) there, so we never noticed the undocumented dependency.

In the long run, `moment` should live in just one place, of course.