UI layout and wording tweaks #94
Attempt a new direction for design and layout:

1. Move "smoke alarms save lives" text to a banner position below logo
and spanning width of screen. The intent of this text is different from
that of the other text in the left hand panel, and it deserves to be set
in its own container. This move also has the nice effect of reducing the
height of the way-too-tall left-hand panel.

2. Remove "Please note that installation requests are not available in
all areas." from the left-hand panel. This text is concerned with
availability, which is addressed in the bottom panel. Now the text in
the left panel flows much more naturally ("...and click 'Submit'" is
followed immediately by "After submitting your report..."). The too-tall
left-hand panel is also further reduced in height, so that now it is
about the same height as the right-hand panel, giving a more balanced
look.

3. Add color images to the left and right columns. This makes the site
much less monochromatic, promotes the brand (the Red Cross logo is
visible in both pictures) and silenty coveys the idea that nice people
will come to your house with all the necessary tools and install a smoke
alarm for you. (Note: the images vanish on smaller screens, where higher
content-density is desirable).

4. Remove the background image entirely. With the color images and
multiple text boxes in place, the background makes the page look too
busy. It might be brought back in a highly washed-out form to add back a
little bit of texture.