Filtering results WIP: #100
Added results to return filtering properly. Considered a WIP
due to challenges working with Backbonejs's routing. Specifically,
once the filter is used to make an XHR request, all subsequent
links within the filter view are considered XHR requests, so the
filtered link does not make it to the browser. If you took the
link and made a plain HTTP request, the link would send a filtered
CSV.

Also, this code added some random instability with Request objects
disappearing from the server requests, requiring a server restart
to get working again. I think this is another Backbone router
interaction, but I'm not sure. In either case, this does not need
to be in master.

I need to learn more about Backbone, its routing and how it works
in order to resolve this issue. In the meantime, this is a WIP

- views/admin/requests/index.jade - Added the ".export" class to
  the "Export CSV" link

- views/admin/requests/index.js - Added a clause that attempted
  to send the CSV type through an XHR request. This didn't work