Clearer UI for dates #103
Add selection of dates via a calender icon that triggers the jquery
datepicker widget. The generic textual input fields into which the
user used to type this information are now hidden but still in the DOM
so that their values (now set by the widget) can be used by the
filtering mechanism.

* smoke-alarm-portal/public/third-party/Farm-Fresh_calendar_view_month.png:
  New file, copied from
  https://commons.wikimedia.org/wiki/File:Farm-Fresh_calendar_view_month.png
  This image is licensed under the Creative Commons Attribution 3.0
  United States license. Attribution is given in
  smoke-alarm-portal/public/third-party/attribution.txt and a copy of
  the license is provided in
  smoke-alarm-portal/public/third-party/cc-license.txt.