Start process of providing static sample request data.
This may be a WIP.  When I try to import the data right now, I get:

    psql:data/requests_static_sample_data.sql:33:
    ERROR: insert or update on table "Requests" violates
    foreign key constraint "Requests_assignedRegion_fkey"
    DETAIL: Key (assignedRegion)=(2991) is not present in
    table "SelectedCounties".

That error might be the result of some not-fully-refreshed database
configuration on my dev box (which, although it has an empty Requests
table, has been through a migration that involved some manual steps),
or it may be that even if we tried this import with a completely new,
up-to-date database it would still not work.  Until I know which, I'm
committing this on the `wip_requests_static_sample_data` branch.

* data/requests_static_sample_data.sql: New file.

* data/fake_request_data.js: Add placeholder where a fix is needed;
  the code here has been broken since commit c338cd606.