Better handling of earliest/latest dates #103
Datepickers now reset to "no date selected" when receiving a date from
the server earlier than the earliest pickable date from the UI (which we
have set at Jan 01, 2000) or later than the latest pickable date (which
is the current date - whatever that happens to be at the time).  These
"out-of-bounds" dates are sent back with the search results when the
user does not specify otherwise (see #114).