Rename columns, #115
Rename `rc_region` to `assigned_rc_region` everywhere.  Rename
`assignedRegion` to `selected_county` everywhere.  Note that
`selectedCounty` (in camelcase) did appear in
public/views/admin/requests/index.js prior to this change.  I think it
was a remnant of some earlier system, because it appeared nowhere else
in the tree.