Add activeRegions table to db: #115
Add a new table to the db that will hold region display names, contact
names, contact emails, and a boolean indicating whether or not it is
active.  The `assigned_rc_region` column in the `Requests` table is a
foreign key to this new table.

I haven't yet changed the region check so that it accounts for whether
or not the region is active (in fact, it doesn't use the new
activeRegions table at all, yet).  I'll do that, and then update
recipients.json to recipients.sql so that I can load it into the
activeRegions table.