Check whether region is active: #115
Check the new activeRegions table to see whether the found region is
active.  Use the table to set displayed region name, contact_name and
contact email address.

This works for active and inactive regions and for zipcodes that are not
in any of the regions.  It fails for invalid zipcodes, but there is
currently a bug on master for those, too.  I'll fix that in the next
commit and remove the remaining reference to recipients.json, which is
in a different file that I didn't alter in this commit.