Remove final references to recipients.json: #115
Remove references to recipients.json and the SelectedCounties table when
filtering on region in the admin area.  As of this commit, the region
codes (and not the displayable names) still show up in the admin area,
but filtering works correctly.

Inactive regions still show up as options in the region dropdown.  We
weren't explicitly asked to remove them, but I think it might be a good
idea.