Show region names in results: #115
Use a reduce() to group a set of Promises that get the display names for
regions from the db.  I needed to return a promise for each request,
holding its display name.  Then, I replace the region code with this
name in each request returned by the query.  The reduce() groups those
together into the results array that is returned to the client.

Thanks very much to @cwebber and @nttaylor for pointers around
JS/callbacks generally and Promises, specifically.