Save serial number in "Requests" table: #69
Create a serial number based on date and number of requests submitted
that date.  Save it to a new "serial" column in the "Requests" table.

Note that there is a possibility at the moment that we could have a
duplicate serial number, because it is constructed from the date and the
count of requests submitted on that date.  If two requests are being
saved at the same time, then they might get the same count of
already-saved requests and then would have the same serial number.

As written in this commit, the serial number is added in an "update."
In the next commit, I might construct and include it in the initial
request to be saved.

I added the serial column to the "Requests" table in this commit, but it
might be better to add it in a migration or via manual SQL.