Catch duplicate serial numbers: #69
Create and save serial number when initially creating request, instead
of updating the request with the serial number afterward.  This allows
us to enforce uniqueness for serial numbers.  Enforcing uniqueness means
that we can "catch" errors on save and assume that the error was a
non-unique serial number.  In the catch function, increment the serial
number and loop back to the original save function as many times as
necessary.

I moved the countRequestsPerDate and padWithZeroes functions, since
they're now called inside getRequestData and need to be declared before
that function.  The lines that had been in updateRequestWithSerial are
now directly in getRequestData, and the serial update function has been
removed.