Return correct response after catch: #69
My catch and save loop was working properly, but I was still getting the
"sorry" page in the browser for valid requests (that should have gotten
the "thank you" page).  Thanks to @nttaylor for pointing out that my
catch() function needed to return a value, instead of just calling the
outer function!