Rewrite serial numbers: #69
Serial numbers will now be constructed by region instead of by date.
They'll be of the form <region code>-<date>-<total number of requests
for that region>.  This means that we need to find the region before
saving the request, because we're using a uniqueness constraint on the
serial number.  I removed the "updateRequestWithRegion" function, since
the region will now be saved with the initial request.

I also moved the code for finding the zip into its own function since it
needs to happen before I save the rest of the request.  Now, we find the
zip, get the county from the address, find the region from the county,
and then save the request.

However, this might break for zipcodes that aren't in any region.  I'll
test and fix that (if necessary) next.