Add new region codes from users.
Add migrations that add the serial column to the Requests table, alter
the region codes to the short human-readable versions, and add the
internal region codes to the db.

Small serial changes, re #69: start request counter at 1 instead of 0.
Change serial to use "XXXX" when state is not found instead of "XXNA."
Presumably that will never happen, since the state is entered from a
dropdown (not as free text), but just in case!  I think "XXXX" shows
more obviously that something is wrong, since "NA" looks very similar to
regular state abbreviations.