Remove option to sort by ID
The *default* is still sorting by the id field in the database.  I
removed this option at @kfogel's suggestion because the ID we're now
displaying is the serial number, and sorting on that is just like
sorting by region.  So, I didn't change any of the sorting on the server
side.  I only removed those options from the client side.