Start #121: Send message via Twilio API
Create a new route and page on the existing smoke alarm app that will
send a TwiML response to the Twilio API.  I'll set this URL
(/path/to/twilio) as the "Request URL" in my Twilio configuration for
testing purposes.  For now, all it will do is send the first response
from our request script.  Next steps are to count the texts sent and
received and send the appropriate responses to them and then actually
save the incoming information in the db.