Continue #121: count views in session
The goal of this change is to count the texts sent to and from a certain
number and use that count to send the correct next text.  It's currently
piggybacking on an existing session that is already set in the app.  I
may need to start a totally different session.  Also, add twilio to our
node module list.