Store the session correctly: #121
Now, actually keep session state between texts from the same phone
number.  As of this commit, the requester should be receiving the
correct messages from the app.  Next I'll need to store the information
received from the requester via SMS as a db request object and include
the correct information from the db in the final response SMS.