Store request data from SMS: #121
Store the incoming data from SMS's as a request object in the database.
Add the parse-address module to parse incoming addresses (we'll need to
test that more, but it works for the simple examples I submitted).

These changes to /views/twilio/index.js make use of a new utilities.js
file.  This file will ultimately replace most of /views/index.js.  I
didn't complete the refactoring in this commit, so right now there is a
fair amount of duplicated code.  I'll clean that up next.

After the cleanup, the next step is to change the contents of our final
text message to use the new serial number and contact phone number from
the db.