Refactor to reuse functions: #121
As promised in 32b1803, remove the functions that are now in
utilities.js from views/index.js.  Small change to the process in that
file to acknowledge that we can create the serial number in a separate
function from getting the rest of the request data.  In utilities.js,
add the state abbreviations and change some references to other
functions to use the "module.exports" prefix.

Unrelatedly, remove the "var client" line from twilio/index.js, because
it was unnecessary and broke functionality.