Save one request per conversation: #121
Several fixes: rename "twilio" dir to "sms", only save each sms
conversation *once* (instead of once per sms!), fix "undefined" in
serial (for cases where the parsing library set "undefined" for the
state, we want to show XXXX, not XXundefined), send email when saving
new SMS request (if it's in an active region).

The main remaining task is to change the content of the final text
depending on whether they're in an eligible region and to include their
serial number.