Send correct "thank you" or "sorry" text: #121
Construct the final text with the serial number and contact information,
and send either a thank you or sorry depending on what zip code was
entered, just like the web site does.

Next: save whether a request came from an sms or site and add a column
for the region contact phone numbers to the db.