Formatting fixes: #121
Pass state to email correctly, remove "undefined" elements from the
street address, and save the request source correctly.

As of this commit, the code for this feature is working.  Next steps are
to procure a real phone number and do more serious testing.  I've only
tested with a Twilio test account.  We need to test with multiple phone
numbers sending texts at the same time.

There are a couple procedural questions and todos about the script, too.
We need to handle the case where they don't send the zipcode.  We may
want to include the source of the request in the email to admins.  I am
currently (in a sense) privileging the option where they send a
different phone number than the one they are texting from.  I may change
the test there to acknowledge that it is likely that they want to use
the same number.

Generally, more testing would be good!