Start #10: example translation
Add a new i18n module and use it to translate one sentence of a
template.  Add the "locales" directory, which is generated by the
module.  I'm not sure if these count as generated files that shouldn't
be versioned or not.  I added them because presumably we'll need to
store the translations that @nttaylor has already done in version
control somewhere.  I commented out a line that allows us to toggle the
language.  Next we'll need to set the language from a cookie.
Presumably the user will click something that sets the cookie, and then
we'll read from there what language to display the page(s) in.