Add Spanish translations #10
Spanish translations are now in place for index.jade, thankyou.jade,
sorry.jade, and layout.jade (which contains the "Javascript is disabled"
message). Note that these translations are present not only for ordinary
text on the front page, but also for form-field placeholders, form
validation error messages, and screen-reader tags. The only text left
untranslated right now is (I think) the value in the html <title> tags
and the header (<h1>) text on the "thank you" and "sorry" pages. I await
resolution of #123 before providing translations for the headers.