Complete Spanish translations #10
Text on the "thank you" and "sorry" pages is now the same in English,
and therefore also in Spanish (see #123). Title tags in the document
head are also translated. This update also contains some changes to the
Enlish locale json file that reflect that all translated strings have
now been invoked in English as well as Spanish (e.g., adding __() around
the "invalid zip code error" and then viewing that error with locale set
to Spanish will add the error text to "es.js", but nothing will happen
to "en.js" until the error is viewed with locale set to English).

Adding the "language switcher" to the UI and using it to set a
locale-specific cookie on the user's browser still remains to be done.