Show request source, check for zipcode: #121
Record the source of the request (SMS or site) in the email that goes to
the admins.  Check whether the requester included the zipcode in their
request text (to do so, add another text to the array and skip it if it
isn't needed).  Add some notes and the start of logic to catch Spanish
requests.  I'll need to do full i18n of this section, as in #10.

Next is to catch the failure if our response text is not received.