Language switcher with minimal styling #10
Implement the language switcher widget in the upper-left corner of the
page.  I deliberately jammed it into the corner as a way to visually
communicate that this control is - obviously - related to the content of
the page and yet at the same time sort of above and beyond it.  It is
unobtrusive but hopefully still noticeable. Selecting a language sets a
cookie on the user's browser so that when the page is refreshed the
language choice remains.

This commit will be quickly followed by one which removes the label
"view this page in" because this label may be unnecessary and at any
rate is written in English and not in a language that someone using the
language switcher would presumably feel more comfortable with!