Different styling for language switcher #10
Remove the label in front of the language switcher and write out the
full names of the language options. Put the options in a box with a
black background and white borders on two sides to improve visibility
while maintaining a feel of the control being connected to the form
while also being "above and beyond" it.

Note that the language switcher will partially cover the logo on very
narrow screen sizes.