i18n for text messages #121
Figure out how to use the i18n-2 module in views/sms/index.js (as
opposed to in a jade template). This is needed in order to send text
messages in the user's chosen language (since the SMS funcionality is
obviously unrelated to any jade template). Also begin the work of
translating the actual messages into Spanish.