Send final text correctly: #121
Send the final text using the "from" number, with Spanish translation.
However, doing this in the "constructFinalText" function makes it more
difficult to increment the cookie properly, meaning that the user
wouldn't be able to start over as they would expect.

Next is to set the cookie properly so that we know when we've sent the
final text, and can start over.  Then, make the string replacement work
in both languages.