Bottom four-panel mosaic for wide screens #122
Wide screens ( > 992 px.) now sport an image panel at the bottom of the
screen with four images showing the installation of a smoke
detector. These panels disappar for screens < 992 px.  Also -- with the
important exception of the language bar -- constrain all visual elements
to the center 80% of the screen when viewing on screens > 992 px. wide.
This makes for more compact text on those screens, which is easier to
read and seems to cause slightly less interference with the main
background image. The width reverts to 100% for screens < 992 px.