Send final text with twiml: #121
Move saveRequest and constructFinalText inside the exports.respond()
function so that they can use the "res" object and successfully send the
text message.  (In order to send a twiml message, Node has to send the
headers and end the response.)

Make sure that we only send the headers once.  In this commit, I assume
that the final text and the headers will be sent from
"constructFinalText()".  I need to do more checking here, since there is
a case where even for the final response we need to send the headers
from the top-level function because we don't have an address and
constructFinalText() will never be called.

Check number of digits correctly -- if there are none, the array is null
and has no length.  In that case, set the number of digits to zero.

Next steps: find/pass the state and county correctly to the final text.
Make sure that the "change language" request has immediate effect
(currently it takes two tries).  General clean up.