Address small translation and script issues: #121
Remove "to contact Red Cross" language in thank you text, because we
don't have contact numbers for regions yet.  Opened #129 for discussion
on whether we'll ever have those.  Fix translation of thank you and
sorry texts so that the full text shows up in the translated language.

Next is to add "statusCallback" to every call to twiml.message(), so
that we can find out why some texts never arrive.