Add a route for the statusCallback: #121
Add the statusCallback parameter when we call twiml.message().  That
takes a URL and posts a status to it, so also create the URL for it to
POST to.

Also, move documentation of constructFinalText inside exports.respond()
so that it is next to the function.  I should have done that in 7a84577.

This is an intermediate commit so that I can test with real texts on the
demo server.