Reorganize texting for #121
Intermediate changes to the SMS saving protocol.  Instead of counting
the texts we've sent in a conversation to determine what kind of
information we need to request and is being received, organize the
information we need to retrieve into a priority list.  Store the
received values in that cookie and send a text based on the next most
important piece of information.

The texts currently have an off-by-one error that I need to debug.  The
request also does not save in this current state.  However, each request
is now correctly differentiated from other requests being processed at
the same time.

Next: fix the off-by-one error in sending texts.  Save request
correctly.  Add error texts.