Add contrasting link background to fix #130
Add dark shading behind the link text on the front page to improve
contrast.  I tried the dark red suggested by @OhMcGoo, but because the
background picture moves from light to dark it was hard to impossible to
find a color that looked good over the whole link.  Adding a shaded
background instead makes the text much more visible.

It is also now bold and I moved the period outside the link.