Allow more than one user to sign up and login: #127
The findOne Sequelize calls were only finding the first user in the
table, so any subsequent user created couldn't log in.

Another issue is that the "login attempt" count doesn't keep track of
whether or not login was successful, so logging in and out successfully
multiple times eventually results in a "too many login attempts" error.
Fixing that will require adding a column to the LoginAttempt table that
records whether or not login was successful, I think.