Show only regions the logged-in user can access: #127
The region dropdown now shows only the regions that the logged-in user
has access to.  The next steps are to (1) make only the requests in
those regions appear on initial page load, (2) for security, *always*
check that the user has access to the region they're searching for (they
could easily alter the querystring) and (3) show the grayed-out
(inaccessible to the logged-in user) regions in the dropdown.