Filter requests by user access: #127
Only show requests that are accessible to the logged-in user.  Tweak UI
to show regions with checkboxes instead of as a select dropdown, so that
the user can view one or more regions that she has access to.  This is
still in an intermediate state with a lot of comments, but seems to be
working for initial tests.

UI improvements to come: show the regions that the user does not have
access to, but as gray (unclickable) checkboxes and text.  On page load,
check the checkboxes next to the regions that have been filtered on (by
default this will be all of them, and show which ones the user searched
on if any).

Bug: no matter what user is logged in, the site thinks that the first
user in the table is logged in.  I suspect this is another "where"
problem like the ones addressed in commit 16c94da.