Show disabled regions: #127
Show users the regions that they are not allowed to filter on, but with
disabled checkboxes and in grayed-out font.

Note that I've changed index.min.css, which is a generated file from
index.less.  Since we had the .min.css file versioned already I've made
the change under version control, but I think we likely want these
generated files out of the tree.