Check/uncheck checkboxes appropriately: #127
On page load, show all checkboxes checked for the regions that the
logged-in user has access to (if any).  When filtering, show the
checkboxes correctly.

Unfortunately, they still don't clear correctly when a user clicks
"clear filters."  That's up next.  Once that's done, I'll construct some
tests and this issue should be complete.